Refactoring nested for-if loop but inappropriate naming
