+++ src/main/java/com/larry/web/HelloController.java
@@ -1,0 +1,7 @@
+package com.larry.web;
+
+import com.larry.webserver.annotations.Controller;
+
+@Controller
+public class HelloController {
+}
