+++ src/main/java/com/larry/webserver/BeanPool.java
@@ -1,0 +1,8 @@
+package com.larry.webserver;
+
+import java.util.Set;
+
+public interface BeanPool {
+
+    Set<Class<?>> getBeans();
+}
