+++ src/main/java/com/larry/webserver/ControllerPool.java
@@ -1,0 +1,17 @@
+package com.larry.webserver;
+
+import java.util.Set;
+
+public class ControllerPool implements BeanPool{
+
+    private BeanFinder finder;
+
+    public ControllerPool(BeanFinder finder) {
+        this.finder = finder;
+    }
+
+    @Override
+    public Set<Class<?>> getBeans() {
+        return finder.find();
+    }
+}
