+++ src/main/java/com/larry/webserver/HttpMethod.java
@@ -1,0 +1,6 @@
+package com.larry.webserver;
+
+public enum HttpMethod {
+
+    GET, POST, PUT, DELETE;
+}
