+++ src/main/java/com/larry/webserver/HttpStatus.java
@@ -1,0 +1,17 @@
+package com.larry.webserver;
+
+public enum HttpStatus {
+
+    OK(200),
+    FOUND(302);
+
+    private int code;
+
+    HttpStatus(int code) {
+        this.code = code;
+    }
+
+    public String toString() {
+        return code + " " + this.name();
+    }
+}
