+++ src/main/java/com/larry/webserver/RequestHandler.java
@@ -1,0 +1,46 @@
+package com.larry.webserver;
+
+import java.io.*;
+import java.net.Socket;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class RequestHandler extends Thread {
+
+    private FrontController frontController;
+
+    private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
+
+    private Socket connection;
+
+    public RequestHandler(Socket connectionSocket, FrontController frontController) {
+        this.connection = connectionSocket;
+        this.frontController = frontController;
+    }
+
+    @Override
+    public void run() {
+        log.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
+                connection.getPort());
+
+        try (InputStream in = connection.getInputStream();
+             OutputStream out = connection.getOutputStream()) {
+
+            BufferedReader br = new BufferedReader(new InputStreamReader(in, "UTF-8"));
+            DataOutputStream dos = new DataOutputStream(out);
+
+            Request request = new Request(br);
+            Response response = frontController.resolveRequest(request);
+            response.send(dos);
+
+        } catch (IOException e) {
+            log.error(e.getMessage());
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        } catch (InstantiationException e) {
+            e.printStackTrace();
+        }
+    }
+
+}
