+++ src/main/java/com/larry/webserver/annotations/LattyFramework.java
@@ -1,0 +1,11 @@
+package com.larry.webserver.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LattyFramework {
+}
