+++ src/main/java/com/larry/webserver/annotations/RequestMapping.java
@@ -1,0 +1,13 @@
+package com.larry.webserver.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RequestMapping {
+    String path() default "/";
+    String method() default "";
+}
