+++ src/main/java/com/larry/webserver/exceptions/ExceptionHandler.java
@@ -1,0 +1,10 @@
+package com.larry.webserver.exceptions;
+
+import java.lang.reflect.InvocationTargetException;
+
+public class ExceptionHandler {
+
+    public static void handle(InvocationTargetException e) {
+        System.out.println(e.getCause());
+    }
+}
