+++ src/main/java/com/larry/webserver/exceptions/UnAuthenticationException.java
@@ -1,0 +1,4 @@
+package com.larry.webserver.exceptions;
+
+public class UnAuthenticationException extends RuntimeException{
+}
