--- src/main/java/web/HelloController.java
@@ -1,7 +1,0 @@
-package web;
-
-import webserver.annotations.Controller;
-
-@Controller
-public class HelloController {
-}
