--- src/main/java/webserver/BeanPool.java
@@ -1,8 +1,0 @@
-package webserver;
-
-import java.util.Set;
-
-public interface BeanPool {
-
-    Set<Class<?>> getControllers();
-}
