--- src/main/java/webserver/BeanResolver.java
@@ -1,5 +1,0 @@
-package webserver;
-
-public interface BeanResolver {
-    Class<?> resolve(Request request);
-}
