--- src/main/java/webserver/HttpMethod.java
@@ -1,6 +1,0 @@
-package webserver;
-
-public enum HttpMethod {
-
-    GET, POST, PUT, DELETE;
-}
