--- src/main/java/webserver/RequestHandler.java
@@ -1,44 +1,0 @@
-package webserver;
-
-import java.io.*;
-import java.lang.reflect.InvocationTargetException;
-import java.net.Socket;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class RequestHandler extends Thread {
-    private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
-
-    private Socket connection;
-
-    public RequestHandler(Socket connectionSocket) {
-        this.connection = connectionSocket;
-    }
-
-    @Override
-    public void run() {
-        log.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
-                connection.getPort());
-
-        try (InputStream in = connection.getInputStream();
-             OutputStream out = connection.getOutputStream()) {
-
-            BufferedReader br = new BufferedReader(new InputStreamReader(in, "UTF-8"));
-            DataOutputStream dos = new DataOutputStream(out);
-
-            Request request = new Request(br);
-            FrontController frontController = new FrontController();
-            Response response = frontController.resolveRequest(request);
-            response.send(dos);
-
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        } catch (InstantiationException e) {
-            e.printStackTrace();
-        }
-    }
-
-}
