--- src/main/java/webserver/WebServer.java
@@ -1,11 +1,0 @@
-package webserver;
-
-import webserver.annotations.LattyFramework;
-
-@LattyFramework
-public class WebServer {
-
-    public static void main(String args[]) {
-        Latty.run(WebServer.class, args);
-    }
-}
