--- src/main/java/webserver/annotations/RequestMapping.java
@@ -1,13 +1,0 @@
-package webserver.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RequestMapping {
-    String path() default "/";
-    String method() default "";
-}
