--- src/main/java/webserver/exceptions/ExceptionHandler.java
@@ -1,10 +1,0 @@
-package webserver.exceptions;
-
-import java.lang.reflect.InvocationTargetException;
-
-public class ExceptionHandler {
-
-    public static void handle(InvocationTargetException e) {
-        System.out.println(e.getCause());
-    }
-}
