--- src/main/java/webserver/exceptions/UnAuthenticationException.java
@@ -1,4 +1,0 @@
-package webserver.exceptions;
-
-public class UnAuthenticationException extends RuntimeException{
-}
